+++ confik/CHANGELOG.md
@@ -2,9 +2,8 @@
 
 ## Unreleased
 
-- Add `SecretOption`, an alternative to `Option` as a `Configuration::Builder` for types that are always secret.
-- Added support for `SecretString` from the [secrecy](https://crates.io/crates/secrecy) crate.
-  - This type is always considered a secret, and can only be loaded from `Source`s which `allow_secrets`.
+- Implement `Configuration` for [`secrecy::SecretString`](https://docs.rs/secrecy/0.8/secrecy/type.SecretString.html). This type is always considered a secret, and can only be loaded from `Source`s which `.allow_secrets()`.
+- Add `SecretOption`, an alternative to `Option` as a `Configuration::Builder` for **types** which are always secret.
 
 ## 0.10.0